Moved components to their own module and files. Split up hangry.rs as well.
