+++ hangry_games/src/gui/components/header.rs
@@ -1,0 +1,11 @@
+use dioxus::prelude::*;
+
+#[component]
+pub fn Header() -> Element {
+    rsx! {
+        h1 {
+            "Hangry Games"
+        }
+    }
+}
+
